Implement an expand feature for job group YAML
We have the parsed scenarios in a hash. Generating a result is as easy
as feeding the data back into the job template format.

The API takes an additional `expand` option to compute the result YAML
after expanding aliases, settings and defaults.

The web UI exposes the feature as a new `Show expanded version` button
which doesn't save the changes.

This also adds additional documentation on the API response to be
clearer about the available fields with different options.

Fixes: https://progress.opensuse.org/issues/56279