Remove wrong error handling code when sending ws messages
It is not possibly to handle errors of the async send() function
immediately. The send() function just returns the transaction. So
the error handling did actually nothing.

Errors when sending ws messages to the worker could only be handled
in the finished event of the ws connection. But so far everything
works well enough without trying to be clever in the error case
so I'm not going to take the effort. The worker will reconnect if
the connection is broken anyways.