Support configurable patterns for asset cleanup
Patterns match the basename and map to a time in days which overrides
the default for untracked assets.

While #2460 removes the hard-coded behavior to never delete repos with
CURRENT in the name, we want to specify when these files will be cleaned
up and it should be longer than the default (30 vs 2 days).

Fixes: https://progress.opensuse.org/issues/59103