Add new icons for every interesting state/result
Created based on base logo which I shifted in position in a previous
commit to make more room for the check mark, the X mark or the state
bubble. Not all states are included to not make it too confusing but a
similar reduction is also done in e.g. templates/tests/infopanel.html.ep
when choosing the color of the panel. For the colors I used the
definitions in assets/stylesheets/openqa_theme.scss

Alternatives for icons could be
https://iconmonstr.com/check-mark-14-svg/ with a round or square border
around the checkmark to make it stand out more.