.circleci: Fix ignored test results of unstable tests within loop
Calling "make test" within a for loop will ignore all results for final
exit code evaluation except the last run which was for
"t/ui/26-jobs_restart.t". This means that all other tests within
"unstable_tests.txt" were effectively ignored. This can for example be
seen in
https://app.circleci.com/jobs/github/os-autoinst/openQA/6274/parallel-runs/0/steps/0-108

This commit fixes this by aborting the for-loop with a failure properly
propagated if any of the "make test" runs fails to succeed within the
specified retries.