Provide investigation hints for failed jobs on request
In the case of a failed job where not carry over is available, for
example in the common case of a failed job following a passed one, the
important question to answer in general is "what changed" which can have
multiple different angles. This commit adds a simple "diff" of the "last
good" job vs. the current "first bad" relying on the file vars.json from
the os-autoinst backend which provides all relevant information for the
most cases.

This commit adds a plain HTML route as well as JSON routes. The route is
called from test details when a user clicks on a tab to gather the
investigation details.

Example output corresponding to the investigation details:

Last good: 99983
Difference of job data:

```
--- t/data/openqa/testresults/00099/00099983-Unicorn-42-pink-x86_64-Build667-K@RainbowPC/vars.json	Tue Dec  3 16:39:15 2019
+++ t/data/openqa/testresults/00099/00099984-Unicorn-42-pink-x86_64-Build668-K@RainbowPC/vars.json	Tue Dec  3 16:39:15 2019
@@ -1,10 +1,10 @@
 {
-   "BUILD" : "667",
+   "BUILD" : "668",
    "DESKTOP" : "textmode",
-   "JOBTOKEN" : "fNBevAKhhOQux_GN",
+   "JOBTOKEN" : "fNBevAKhhOQux_GO",
    "NEEDLES_GIT_HASH" : "3aee70f5f526175469aa189b0c4f820de8f06ae1",
-   "QEMUPORT" : 20001,
+   "QEMUPORT" : 20002,
    "TEST_GIT_HASH" : "c65c426a2678223fad4ae7d4ebbe95817e5c725d",
    "TEXTMODE" : 1,
-   "WORKER_INSTANCE" : 1
+   "WORKER_INSTANCE" : 2
 }
```

https://progress.opensuse.org/issues/39719