openqa-clone-custom-git-refspec: Add support for comma-separated job list
This adds support to specify a comma-separated list of job URLs instead of a
single one as second command line parameter. The clone with according
parameters is then conducted for each specified list.

Also this adds a "markdown" mode especially useful when a list of jobs is
specified and one wants to copy the command output into a github pull request
in markdown list format.

Example output:
$ env MARKDOWN=1 openqa-clone-custom-git-refspec \
    https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/7084 \
    https://openqa.opensuse.org/tests/1105660,\
    https://openqa.opensuse.org/tests/1105386,\
    https://openqa.suse.de/tests/3676702

* [opensuse-Tumbleweed-DVD-x86_64-Build20191206-kde-wayland@64bit_virtio](https://openqa.opensuse.org/t1105940)
* [opensuse-Tumbleweed-DVD-x86_64-Build20191206-extra_tests_on_kde@64bit](https://openqa.opensuse.org/t1105941)
* [sle-12-SP3-Server-DVD-Updates-x86_64-Build20191207-1-qam-minimal+base@64bit](https://openqa.suse.de/t3677892)

https://progress.opensuse.org/issues/58184