job investigation: Give explanation on no test changes
Looking up the git log diff between "last good" and "first bad" of a
scenario history can yield an empty result meaning that there are no
changes, i.e. both git hashes most likely match. If this is the case we
can point the reviewer to the likely hypothesis that test changes are
not the cause of a job failure.

Related progress issue: https://progress.opensuse.org/issues/19720