Optimize job module insertion with raw SQL conflict handling
Same as previously for screenshots where we fixed the same error in both
locations where a DB insertion is done already in
https://github.com/os-autoinst/openQA/pull/1792 and
https://github.com/os-autoinst/openQA/pull/2597 we should also prevent
expected conflicts to cause an error entry in postgres logs about job
modules. This is a followup to ecb7d1d50 where the unique constraint for
job modules was introduced.

Related progress issue: https://progress.opensuse.org/issues/60923