Move webapi templates from "templates/" to "templates/webapi/"
OpenQA started as a single application, so one templates directory made
sense. But now we have multiple micro services that all inherit this
directory by default.

This causes unintended conflicts when for example an exception happens
in one of those services, and the framework exception handler loads
"templates/exception.html.ep". Which was designed for a different
service (webapi in this case). The exception template uses a layout
that expects certain webapi specific values in the stash. And if they
are not present it will die itself with a different exception.

So the correct solution is to have micro service specific templates
directories, starting with "templates/webapi/". And if other services
want to use templates too they can use "templates/websockets/" and the
like.