fetchneedles: Fix "no tracking information for the current branch" on self-repair
2aadd3e56 was probably incomplete in the sense that we recreate the master
branch but never set the upstream information causing the problem:

```
There is no tracking information for the current branch.
Please specify which branch you want to rebase against.
```

To prevent that we can set the upstream branch when recreating a branch.
Additionally it is a safer approach if we specify explicitly which branch to
fetch and rebase against as we already have the branch/needles_branch as
configuration parameters supplied.

Tested manually:

* Test 1: needles repo on valid branch -> fetch+rebase with explicit branch
  works fine

* Test 2: `git co HEAD^^^^` to detach from master -> fetchneedles calls
  self-repair including `git branch --set-upstream-to=origin/master master`

Related progress issue: https://progress.opensuse.org/issues/63133