Redirect to remote code location for source view when using CASEDIR
The source code view always referenced the current state of test code in
a local path rather than the code that was actually used during test
execution. Any use of remote code locations would in many cases be
incorrectly mapped to local paths which might not even exist.

This commit extends the behaviour to redirect to remote code locations in
case if the variable CASEDIR is specified with a URL pointing to a remote
location.

Related progress issue: https://progress.opensuse.org/issues/54965