Do not capture and store all job output as a note
In theory this was a useful feature. But unfortunately in practice it
did not work out so well. Instead it got in the way of debugging tests
(by hiding job output), and in production some jobs generated so much
output that it would overwhelm the database. And we ended up with
workarounds like redirecting gru logs into a file just to avoid this.