Have openqa-clone-custom-git-refspec parse PR for test URLs
Given a PR on GitHub, having one or more tests in the description should
be sufficient to run them without having to specify them on the command line.

This is a prerequisite to automated test runs for PRs in CI.

Fixes: [poo#63712](https://progress.opensuse.org/issues/63712)

Note: This is a revision of #2618 with additional tests and better
handling of text and multiple URLs with no markers.