Bring back elapsed time in CI tests for easier identification of time-hogs
With travis CI we had the timer output. I do not think the change to
remove that was done with an intention but rather by mistake so let's
add back the timer output which can help to identify which tests add
significantly to overall elapsed time. This can help to identify
performance regressions as well as help identify where test collections
can be split to reduce overall runtime.