t: Fix grutasks taking excessively wrong while gathering coverage data
https://app.circleci.com/jobs/github/os-autoinst/openQA/15590 showed that
t/14-grutasks.t was taking 519s (!) to complete , This was not confirmed when
running normal tests locally. In my not so fast environment the test repeatedly
takes about 16s (only). This could explain why "t" seems to take comparably
longer in the recent past. This appears to be a performance regression. I
recorded profiling data locally with

```
PERL5LIB=lib:$PERL5LIB perl -d:NYTProf t/14-grutasks.t
nytprofhtml --open
```

This showed that the biggest time impact is in

```
130 1   1   134s    134s                       Minion::Job::CORE:waitpid (opcode)
```

which I understand as that we simply wait for 130 minion jobs to
succeed. The hypothesis was that when trying to collect coverage data we
call code for every single minion job which could also accumulate to a
very long time. A temporary change to patch out the coverage data
collection proved that the test runtime was reduced again significantly.

This commit changes the test to only gather coverage data for just one out of
100 minion jobs which all execute the same task anyway. This is implemented by
passing a "note" with all but the first minion job in the TTL subtest to skip
coverage data collection.

Related progress issue: https://progress.opensuse.org/issues/63985