Fix script/retry for STABILITY_TEST preventing false-negatives
When the command within "run_once" fails we must not only break the loop
but also preserve the exit code. This was observed when trying out
STABILITY_TEST in more cases which is not affecting our current circle
CI based tests as they do not use STABILITY_TEST anywhere in the current
state.