Makefile: Prevent unexpected self-test run time increases with timeouts
As https://progress.opensuse.org/issues/63985 showed it can happen that
tests increase in the run time they take without us noticing at the time
when a change is introduced. Would be good if we can prevent tests
taking significantly longer than they did before.

With this commit significant increases of test run time can be caught
when running tests locally on a global level as well as on the level of
test groups which is also effective for circle CI. Current timeout
values are a conservative limit rounded up from current run times on
circle CI which seems to provide reliable and reproducible test run
times.

The timeout value is scaled with the number of applied retries as well.

Related progress issue: https://progress.opensuse.org/issues/64216