docs: Move growing "testing" section out of "commit rules"
We started with a simple and small list in "rules for commit". Within
that list we had a growing part about how to conduct local tests. In
another location within the document there are already mentions for test
setup and execution which should be merged.

This commit moves the test instructions to its own section together with
existing mentions of test setup. Additionally the description of manual
setup is streamlined.