Remove untested and probably unused iChain authentification
6b6002da1 added iChain, in github PR
https://github.com/os-autoinst/openQA/pull/160
and progress ticket https://progress.opensuse.org/issues/1729. When
trying to cover authentication modules with better tests I realized that
there is already a typo in one of the request parameters. I doubt iChain
authentication ever worked or would still work. I assume it is not used
at all so we are better off removing it from the current code.