Separate all "scripts" to be packaged from development "tools"
This also simplifies installation rules as we do not need to mention
individual files to be deleted from user facing packages but can rely on
folder paths.

Another benefit is that we are now using the same setup as in the main
backend "os-autoinst" so that for example common contributors of both
are not confused by "script/tidy" in openQA vs. "tools/tidy" in
os-autoinst.