Fix load_templates --clean to wipe more than one entry per table
1400d8a8dd inverted the logic here - instead of exiting the loop
on a *failure* we now exit it on *success*. So each time --clean
is run it will only wipe one entry per table. This obviously
breaks things badly (the clean phase doesn't clean things right
at all, so the subsequent load phase fails because we try to add
entries that already exist, and you wind up with a mess).

This flips the logic back, and adds a test.

Signed-off-by: Adam Williamson <awilliam@redhat.com>