t: Try to stabilize t/ui/01-list.t
We do not really expect "animations" in this step but it might be the
data table is built not immediately after user actions to provide a more
responsive user interface.

I failed to reliably reproduce the problem so the fix is without actual
verification.

Related progress issue: https://progress.opensuse.org/issues/59043