t: Ensure all test code is executed in every environment
Some part of the code in t/full-stack.t was only executed if the
condition `!$ENV{MOJO_LOG_LEVEL} || $ENV{MOJO_LOG_LEVEL} =~
/DEBUG|INFO/i` is true which apparently was not the case in circleCI
causing relevant test code not never be executed in circleCI but when
executing the full stack test locally the test could fail since the test
is in need of adaption. We should not have conditional test code that is
never executed in our CI environment. In this case we should be able to
rely on the worker being correctly started so that any other configured
log level selection has no effect. To fix the tests I adapted the
assertions accordingly.