t: Prevent use of occupied port in all full-stack/scalability tests
Sporadically tests can fail with "Connection refused" when already
occupied ports have been selected for services in tests as we only
looked for a single free port and tried to use the ports next to it
without checking if they are actually free.

This commit is inspired from the idea in the scalability test extended
to all services and used in all relevant full-stack and scalability
tests. By mocking the function "service_port" from OpenQA::Utils
whenever the new function "mock_service_ports" is called we can provide
a consistent but dynamically defined set of ports that should all be
free to use during the course of each test.

Related progress issue: https://progress.opensuse.org/issues/59043