Define custom log color selection compatible with reverse video terminals
The default selection of Mojo::Log::Colored is not compatible with a
reverse video terminal, i.e. black font on white background. This commit
selects a compromise selection that works for both black-on-white as
well as white-on-black. Also the selection is more similar to what we
had originally in https://github.com/os-autoinst/os-autoinst/pull/876
before support for color has been removed.

The selection of colors has no effect if Mojo::Log::Colored is not
available.