Allow to configure worker connect interval using env variable
Adding new environment variable configuration
"OPENQA_WORKER_CONNECT_INTERVAL" which can be set in the environment and
is also set in tests to prevent unnecessary waiting for any non-existant
or slow test webui process.

See
https://app.circleci.com/pipelines/github/os-autoinst/openQA/3092/workflows/7f45c7f4-44ca-40c4-9629-2c8342e23fee/jobs/29471/steps
for a related failed test within circleci that was aborted due to no
text printed for more than 10 minutes. In an example like this we can at least
fail the test earlier with a more meaningful feedback in case the webui
process is not existant (anymore) at all to connect to.

Related progress issue: https://progress.opensuse.org/issues/59043