Prevent schedule from being reread unnecessarily
This can happen if a test is finished and the new one hasn't started.
(Reproduce by adding a `sleep 10` in basetest.pm after `$self->done()`.

The backend will report an empty `$running_test`, and in this case
`$current_test_module` should keep its last value.

It will also reduce payload of the status posts, because it previously posted
the status of *all* tests when `$current_test_module` was empty.

This also fixes the issue that results of short tests < 10s are only
reported at the end of the job.

Related Issues:
* https://progress.opensuse.org/issues/58826
* https://progress.opensuse.org/issues/58700
* https://progress.opensuse.org/issues/67972