Revert "Increase version requirement for some dependencies"
Reverts os-autoinst/openQA#3189

Currently we can not fulfill the requirements in our common
environments, e.g.
https://openqa.opensuse.org/tests/1311407#step/openqa_webui/21 shows
based on latest Tumbleweed . So I would say the PR was correct but
should not yet have been merged. Also
https://build.opensuse.org/package/show/openSUSE:Factory/perl-Minion-Backend-SQLite
is not linked to
https://build.opensuse.org/project/show/devel:openQA:Leap:15.1 neither
to https://build.opensuse.org/project/show/devel:openQA:Leap:15.2 . I
understand that the new upstream versions provide helpful bugfixes but I
do not see them as hard requirements. As soon as newer package versions
are present we benefit from them but they are not hard requirements.
I recommend to revert this PR.