deps: Fix nightly dependency bot job
Before, the job was running a docker in the already running docker.
CircleCI doesn't allow mounting volumes in that scenario, that's why the
`make update-deps` and `tools/tidy` never worked.
We just didn't see that because we never had a perltidy update since that
code was added.

This commit simply runs the actual job in the already running docker.

It only needs to install `hub` to make the actual PR (in case of changes).

So we don't need an extra image for the bot, but I left the file in place so
far:
    docker/devel:openQA:ci/dependency_bot/Dockerfile

I tested it in a branch in the origin repo to see if it created a PR: #3248