Make "last good" a link to a job instead of plain job ID
https://progress.opensuse.org/issues/19720

If a valid last_good job ID is present in the investigation result
set, then it is replaced by html code <a> to link directly to this job.

In order to tell to the JS that this is a link, backend adds
the type of this variable. Using ':' to split <name>[:<type>].
In this case the type is link. The key will be now 'last_good:link'

In JS detects if a type has being defined or not. If is link then
creates a link with the information. In other cases, now only
traditional use, then creates a <pre> element.

This allows to add typed information like links to files,
tables, but also colorize the values is needed.