spec: Save time in package build tests
This commit changes the spec file to not run static source code checks
in package build tests as they are less (or not all) useful in
repository specific test runs. So we can save runtime here.

https://build.opensuse.org/package/live_build_log/devel:openQA:TestGithub:PR-3270/openQA/openSUSE_Tumbleweed/x86_64
shows a runtime of 750s with the changes of this commit in comparison to
https://build.opensuse.org/package/live_build_log/devel:openQA/openQA/openSUSE_Tumbleweed/x86_64
with a runtime of 1000s so a saved time of 210s.