Show changes between packages of openQA worker machines in investigation
https://progress.opensuse.org/issues/19720
https://progress.opensuse.org/issues/69088

1. Extract and stores the package list
2. Compare the current job list against the last good job list

The solution within openQA is agnostic to the operating system
of the openQA worker. To solve that the command to list the
packages is configured in the worker.ini, for instance:

PACKAGES_CMD = rpm -qa

The worker_packages.txt files are checheck before to compare. If current
job or last good job doesn't have this file shows an alert.

The worker_packages.txt is created when the worker starts

Add test to check the diff generation of packages
Add a test that the execution of the packages list command works

Other considerations when the package list command is called:
- The execution error code is checked
- The result file is checked to have more than 0 bytes of length