`make test-with-database`: add -I$(PWD)/t/lib back to PERL5OPT
As discussed in https://github.com/os-autoinst/openQA/pull/3194
while changing `-I$(PWD)/t/lib` to `-It/lib` made things work in
Docker, it broke them in Fedora's build system. We really don't
know why and it's a bit of a mystery, but it seems easier to just
have both than try and figure out the causes.

Signed-off-by: Adam Williamson <awilliam@redhat.com>