Make localhost uploads from worker to webui faster
When the LOCAL_UPLOAD=1 setting is present in the workers.ini, the
worker will check if the webui is on localhost/127.0.0.1 and then
copy assets directly on the file system instead of uploading them
via HTTP. In benchmarks uploading a 200MB asset (20 chunks) this
has been about 27 times faster than before.

If the feature proves stable in production it might be worth
removing the LOCAL_UPLOAD setting in the future and relying only
on host detection. We might need to add an environment variable
instead however, to ensure full stack tests stay full stack and
don't end up using only this optimization.