Simplify insertion of preview container within test details page
Inserting the preview container right after the thumbnail element just
works. It works also when the thumbnails are wrapped into a second row.

I don't know what the point of the previous code was but it just looks
more complicated than it needs to be and doesn't work when an external
parser result is present in the middle of a test module.