Surface errors from failed test tries
The retry script can re-run a test that fails at the cost
of losing the results (since we do not store them anywhere).
We can get failures from TAP output (even if the testplan broken).
This relies on the hardness dump which we use in CI. It can
be enabled locally by setting the variable if desired.

See: https://progress.opensuse.org/issues/66664