Avoid worker assumed to be offline from interfering when it comes back after all
The worker's status updates are already rejected by the web UI when the ID
specified by the worker does not match the ID of the actually assigned
worker. This change applies the same check to the API function to mark a
job as done.

So now, when a worker is assumed offline the job can be re-scheduled and
assigned to a different worker without running into problems when the
worker assumed to be offline comes back after all.

See https://progress.opensuse.org/issues/64520