Allow cancelling assigned jobs like scheduled jobs and prevent restart
So far it was possible to restart assigned jobs but that left the old job
as-is (stuck in assigned). Now the assigned job can be cancelled like
scheduled jobs.

If the worker has just started working on the job after all it will stop
the execution but the resulting API errors will not interfere because the
that is already covered by the previous commit.

Cancelling a pending directly chained children works as well. The worker
will still execute the current job in the direct chain but skip the
cancelled child (and its children).

See https://progress.opensuse.org/issues/64520