t: Remove obsolete test job database entry in all initializations
Since a very long time in our initialization of test database entries we
updated the timestamp of one job, 99937. This was likely necessary for a
specific test and definitely not needed in all test modules that use the
same test database. Apparently the timestamp update is not even
necessary anymore in any current test module so this commit is removing
the according code completely.