t: Add scaling parameters for OpenQA::Test::TimeLimit
We know from experience that tests within circleCI can run slower so
this commit applies a scaling of the test timeout within the
OpenQA::Test::TimeLimit module with a default parameter and the option
to override the value using an environment variable. Also when running
with Devel::Cover to conduct a code coverage analysis the test runtime
can increase significantly so an additional parameter for scaling is
introduced here.

Instead of scaling up the time in each test module we should apply
sensible global scaling parameters and keep the limits in the test
modules specific and based on reliable run time analysis without the
scaling implied.

Related progress issue: https://progress.opensuse.org/issues/71845