Extend web UI docker setup to use haproxy/docker-compose
Fix the dockerfile and adds a docker-compose file and the
config files in order to have an HA web UI environment for
testing and development purposes. This deploy the following containers:

- There are 2 web UI container by default.
- There is a haproxy load balancing service to balance
the load of the web UI. This exposes the port 9526
- The scheduler, websockets, gru and livehandler are
independent containers
- The DB is a postgresql container

Some considerations:

The number of replicas for achieve the HA of the web UI
could be setup with `export OPENQA_WEBUI_REPLICAS=3`.
You could also edit the .env file to change the default value

You can enable the SSL for the load balancing following
the instructions of the readme.md

The haproxy is the load balancing solution for this example,
but could be others like nginx or apache.
Also, haproxy could has more sophisticated configuration,
for instance to checking the health of the web UI containers

Remove Fedora guide to launch web UI via Docker