Extend web UI docker setup to use nginx/docker-compose
Fix the Dockerfile and add a docker-compose file as well as config
files in order to have an HA web UI environment for testing and
development purposes. This deploys the following containers:

- There are 2 web UI container by default.
- There is a nginx load balancing service for the web UI.
This exposes port 9526.
- The scheduler, websockets, gru and livehandler
are independent containers.
- The DB is a postgresql container.

Some considerations:

The number of replicas to achieve the HA of the web UI can be setup
with export OPENQA_WEBUI_REPLICAS=3. It is also possible to edit
the .env file to change the default value

SSL for the load balancing can be enabled following the instructions
in Readme.md.

https://progress.opensuse.org/issues/69355.

This is an alternative to #3431 which uses haproxy instead of nginx.
Websockets and livehandler can be integrated on the same port as
the web UI. For instance /api/v1/ws for the websockets. And this
solves a communication problem with the workers