Ensure all scripts exit non-zero in case of invalid parameters
To prevent incorrect setup of services or scripts in automation we
should ensure that all scripts return a non-zero exit code signalling a
failure.

This commit adds corresponding handling of parameter parsing and
forwarding of the corresponding exit code in all scripts where not
already present. Also adding a corresponding test in t/44-scripts.t.
This increases the runtime especially due to the costly startup of
services but I consider it still beneficial to cover.

Related progress issue: https://progress.opensuse.org/issues/68167