Scale timeout in t/01-test-utilities.t to prevent failures observed on OBS
We've seen the following test failure on OBS (see
https://progress.opensuse.org/issues/73162):

```
[  859s]     # PID of test-process: 5623
[  859s]     ok 1 - warning logged
[  859s]     not ok 2 - test would have failed
[  859s]     # PID of test-process: 5666
[  859s]     ok 3 - message of test failure
[  859s]     ok 4 - exit code of test failure is non-zero
[  859s]     not ok 5 - manual termination via stop_service does not trigger _fail_and_exit
```

The test is actually just running into a timeout. The last failure can be
reproduced locally by adding a sleep at the right place to run there into
the timeout as well. The first failure can also be reproduced by locally
by removing the first sleep.

This PR adds better log message and applies the same scaling used for the
overall timeout to the other timeouts within the test.